Maj du 23/03/2018 16h41

+++ GestionMagasin-ejb/src/java/facades/FournisseurFacade.java
@@ -40,7 +40,7 @@ public class FournisseurFacade extends AbstractFacade<Fournisseur> implements Fo
     }
   
     @Override
-    public Fournisseur rechercherFournisseurParId(int id)
+    public Fournisseur rechercherFournisseurParId(long id)
     {
          Fournisseur result;
